fix(data-audit): add type conversion for Hibernate 7 native query result compatibility
