+++ jobs/kogito-addons-springboot-embedded-jobs/src/main/java/org/kie/kogito/app/jobs/springboot/SpringbootJobServiceConfiguration.java
@@ -32,15 +32,18 @@ import org.kie.kogito.services.uow.DefaultUnitOfWorkManager;
 import org.kie.kogito.uow.UnitOfWorkManager;
 import org.kie.kogito.usertask.UserTasks;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.config.BeanPostProcessor;
 import org.springframework.boot.SpringBootConfiguration;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
-import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
 
+// Note: The Hibernate 7 + Spring ORM 6.2 EntityManagerFactory workaround (BeanPostProcessor for
+// setEntityManagerFactoryInterface) is intentionally NOT in this class. This module does not have
+// spring-orm as a dependency since it supports both JPA and non-JPA configurations.
+// The workaround is located in SpringbootJPAJobStoreConfiguration in the JPA-specific module:
+// kogito-addons-springboot-embedded-jobs-jpa
 @SpringBootConfiguration
 @EnableAutoConfiguration
 @ComponentScan
@@ -91,20 +94,4 @@ public class SpringbootJobServiceConfiguration {
         return new UserTaskInstanceJobExecutor(userTasks, unitOfWorkManager);
     }
 
-    // Hibernate 7 + Spring ORM 6.2 workaround: Hibernate 7's SessionFactory.getSchemaManager()
-    // returns org.hibernate.relational.SchemaManager, conflicting with JPA 3.2's
-    // EntityManagerFactory.getSchemaManager() returning jakarta.persistence.SchemaManager.
-    // Force plain JPA interface to avoid JDK Proxy incompatible return type error.
-    @Bean
-    public static BeanPostProcessor entityManagerFactoryInterfacePostProcessor() {
-        return new BeanPostProcessor() {
-            @Override
-            public Object postProcessBeforeInitialization(Object bean, String beanName) {
-                if (bean instanceof LocalContainerEntityManagerFactoryBean emfb) {
-                    emfb.setEntityManagerFactoryInterface(jakarta.persistence.EntityManagerFactory.class);
-                }
-                return bean;
-            }
-        };
-    }
 }
